goes to next dialog when option is made
need to do more testing, if there is a better way to format how the index is changed to prevent messing up it should be done. (instead of making it foolproof add a debug line to show that its been done incorrectly?) it would be better to add things to groups to ensure the buttons getting freed are the correct ones, this would only be a problem if ButtonContainer (meant for temporary option buttons) has more buttons added manually that aren't meant to be freed.